+++ HLTrigger/HLTanalyzers/test/RateEff/OHltRatePrinter.h
@@ -16,28 +16,28 @@ class OHltRatePrinter
 {
 public:
 
-   static const float fTwo=2.;
+   static const double fTwo=2.;
 
    OHltRatePrinter() {}
 
    virtual ~OHltRatePrinter() {}
 
    void SetupAll(
-         std::vector<float> Rate,
-         std::vector<float> RateErr,
-         std::vector<float> spureRate,
-         std::vector<float> spureRateErr,
-         std::vector<float> pureRate,
-         std::vector<float> pureRateErr,
-         std::vector< std::vector<float> >coMa,
-         std::vector< std::vector<float> > RatePerLS,
+         std::vector<double> Rate,
+         std::vector<double> RateErr,
+         std::vector<double> spureRate,
+         std::vector<double> spureRateErr,
+         std::vector<double> pureRate,
+         std::vector<double> pureRateErr,
+         std::vector< std::vector<double> >coMa,
+         std::vector< std::vector<double> > RatePerLS,
          std::vector<int> tRunID,
          std::vector<int> tLumiSection,
-         std::vector<float> tTotalRatePerLS,
+         std::vector<double> tTotalRatePerLS,
          std::vector< std::vector<int> > tRefPrescalePerLS,
          std::vector< std::vector<int> > tRefL1PrescalePerLS,
-         std::vector<float> averageRefPrescaleHLT,
-         std::vector<float> averageRefPrescaleL1,
+         std::vector<double> averageRefPrescaleHLT,
+         std::vector<double> averageRefPrescaleL1,
 	 std::vector< std::vector<int> > CountPerLS,
 	 std::vector<int> tTotalCountPerLS,
 	 std::vector<double> tLumiPerLS);
@@ -81,24 +81,24 @@ public:
 
    int ivecMin(std::vector<int> ivec);
    
-   std::vector<float> Rate;
-   std::vector<float> RateErr;
-   std::vector<float> spureRate;
-   std::vector<float> spureRateErr;
-   std::vector<float> pureRate;
-   std::vector<float> pureRateErr;
-   std::vector< std::vector<float> > coMa;
+   std::vector<double> Rate;
+   std::vector<double> RateErr;
+   std::vector<double> spureRate;
+   std::vector<double> spureRateErr;
+   std::vector<double> pureRate;
+   std::vector<double> pureRateErr;
+   std::vector< std::vector<double> > coMa;
    std::vector< std::vector<int> > CountPerLS;
    std::vector<int> totalCountPerLS;
 
-   std::vector< std::vector<float> > RatePerLS;
-   std::vector<float> totalRatePerLS;
+   std::vector< std::vector<double> > RatePerLS;
+   std::vector<double> totalRatePerLS;
    std::vector< std::vector<int> > prescaleRefPerLS;
    std::vector< std::vector<int> > prescaleL1RefPerLS;
    std::vector<int> runID;
    std::vector<int> lumiSection;
-   std::vector<float> averageRefPrescaleHLT;
-   std::vector<float> averageRefPrescaleL1;
+   std::vector<double> averageRefPrescaleHLT;
+   std::vector<double> averageRefPrescaleL1;
    std::vector<double> LumiPerLS;
 };
